kraken: fix misunderstanding regarding pair altnames
I was trying to get the ticker for an altname (eg. XBTUSD) and then I
was hardcoding a conversion from an altname to the proper name (eg.
XXBTZUSD). With this patch we use the proper name by default which is
what the api always returns. Optionally a user can make a request with
the altname but will then still have to specify the proper name as a
second argument so we can parse the json returned by the API.